refactor(api): extract tool-profile handlers from system.rs (#3749 11/N)
Move /profiles and /profiles/{name} handlers into a new routes::profiles
module. Public paths and JSON shapes are unchanged; the new router is
merged into v1 alongside the existing system router.

Refs #3749
