+++ crates/librefang-api/src/routes/mod.rs
@@ -27,6 +27,7 @@ pub mod media;
 pub mod memory;
 pub mod network;
 pub mod plugins;
+pub mod profiles;
 pub mod prompts;
 pub mod providers;
 pub mod skills;
@@ -59,6 +60,7 @@ pub use media::*;
 pub use memory::*;
 pub use network::*;
 pub use plugins::*;
+pub use profiles::*;
 pub use providers::*;
 pub use skills::*;
 pub use system::*;
