+++ crates/librefang-api/src/server.rs
@@ -56,6 +56,7 @@ fn api_v1_routes() -> Router<Arc<AppState>> {
         .merge(routes::skills::router())
         .merge(routes::network::router())
         .merge(routes::plugins::router())
+        .merge(routes::profiles::router())
         .merge(routes::providers::router())
         .merge(routes::budget::router())
         .merge(routes::auto_dream::router())
